pyproject.toml
[tool.poetry]
name = "proof-repair"
package-mode = false
version = "0.1.0"
description = "Tool to automatically fix Dafny proofs"
authors = ["emugnier <emugnier@ucsd.edu>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
openai = "^1.12.0"
ruff = "^0.1.0"
pyyaml = "^6.0.1"
pandas = "^2.1.3"
matplotlib = "^3.8.1"
tiktoken = "^0.5.2"
guidance = "^0.1.10"
gspread = "^6.0.2"
multiprocess = "^0.70.16"
numpy = "^1.26.4"
scipy = "^1.12.0"
scikit-learn = "^1.4.1.post1"


[tool.poetry.group.dev.dependencies]
black = {extras = ["jupyter"], version = "^23.11.0"}
jupyter = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
ignore = [
  "E402",
  "E731"
]
exclude = [
  ".ipynb_checkpoints",
  ".pre-commit-config.yaml",
]

[tool.black]
exclude = '''
/(
  \.git
| \.hg
| \.mypy_cache
| \.tox
| \.venv
| _build
| buck-out
| build
| dist
| .*ipynb_checkpoints.*
| .pre-commit-config.yaml
)/
'''

[tool.poetry.scripts]
pythonpath = "tests"
test = 'tests:test_main.__main__'
